Fixed weapon selection, made the ybot turn ragdoll, added blaster, etc(READ description)
This is my on and off progress doing trial and error from today(Wednesday)

I've excluded a Unity package I got from the assest store containing a bunch of fbx files for animations. It can be found at https://assetstore.unity.com/packages/3d/animations/raw-mocap-data-for-mecanim-5330

I tried toying with the ybot spawning with proper animations but it was troublesome and figure someone else could look at it. Just need to modify the animation controller on the ybot and the falldownonclick c# script if interested.

The dummy does not spawn back until there is a no input from the user. Having the buddy spawn in the same location as where it lands is a real task...for reference on how to do it look at the ragdollhelper c# script in the 'Scripts' folder.

If you're going to use an animation for the ybot it must be set to humanoid and not generic.

Ignore the cannonball files I was trying to toy with a projectile. Speaking of someone needs to make projectiles that move the ybot while in ragdoll state.

Notice that my Money system does not work maybe I removed a script from a certain object...BE AWARE and if you see the fix do it and notify others.